requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
requests==2.26.0